// dev/Reactjs/newtech/src/utils/enums.tsx
export enum COLORS {
    DEFAULT_ICON_COLOR = "rgba(0, 0, 0, 0.54)",
    PRIMARY = "#3f51b5",
    PRIMARY_DARK = "#2d3a87"
}

export enum Routes {
    Screen1 = "/",
    Screen2 = "/Screen2"
}

export enum ItemTypes {
    PENDING = 'In Progress',
    DONE = 'Done',
    TODO = 'To Do',
    BOX = 'box'
}